+++ queryset_sequence/__init__.py
@@ -607,13 +607,19 @@ class QuerySetSequence(ComparatorMixin):
         raise NotImplementedError()
 
     def defer(self, *fields):
-        raise NotImplementedError()
+        clone = self._clone()
+        clone._querysets = [qs.defer(*fields) for qs in self._querysets]
+        return clone
 
     def only(self, *fields):
-        raise NotImplementedError()
+        clone = self._clone()
+        clone._querysets = [qs.only(*fields) for qs in self._querysets]
+        return clone
 
     def using(self, alias):
-        raise NotImplementedError()
+        clone = self._clone()
+        clone._querysets = [qs.using(alias) for qs in self._querysets]
+        return clone
 
     def select_for_update(self):
         raise NotImplementedError()