+++ CHANGELOG.rst
@@ -8,6 +8,9 @@ next (2018-xx-xx)
 
 * [Enhancement] Support ``first()`` and ``last()`` methods.
 * [Enhancement] Support the ``&`` and ``|`` operators.
+* [Enhancement] Support ``defer()`` and ``only()`` methods.
+* [Enhancement] Support calling ``using()`` to switch databases for an entire
+  ``QuerySetSequence``.
 * [Bugfix] Raise ``NotImplementedError`` on unimplemented methods. This fixes a
   regression introduced in 0.9.